requirements.txt
numpy
pandas
pysam
tqdm
biopython
recordclass
jellyfish